+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasetop.py
@@ -78,5 +78,5 @@ class RetrieveDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-08-26T09:10:19.431Z")
+    updated: Optional[datetime] = parse_datetime("2025-08-26T12:57:17.526Z")
     r"""The date and time the resource was last updated"""